Conflicts:
	server/api/src/main/java/me/eddiep/ghost/game/match/item/JamItem.java

+++ server/api/src/main/java/me/eddiep/ghost/game/match/entities/TypeableEntity.java
@@ -1,5 +1,5 @@
 package me.eddiep.ghost.game.match.entities;
 
 public interface TypeableEntity extends Entity {
-    byte getType();
+    short getType();
 }
