+++ server/api/src/main/java/me/eddiep/ghost/game/match/entities/ability/CircleEntity.java
@@ -24,7 +24,7 @@ public class CircleEntity extends BaseEntity implements TypeableEntity {
     }
 
     @Override
-    public byte getType() {
+    public short getType() {
         return 4;
     }
