+++ server/api/src/main/java/me/eddiep/ghost/game/match/entities/items/FireRateItemEntity.java
@@ -9,7 +9,7 @@ public class FireRateItemEntity extends ItemEntity {
     }
 
     @Override
-    public byte getType() {
+    public short getType() {
         return 16;
     }
 }
